[Angiogenesis] #2636 Modified map structure to make slightly faster. Changed LinearSystem solver in SimpleFlowSolver to use a direct method which is faster and more accurate than the iterative method being used before. This required making a change to the LinearSystem class to take different preconditioner and ksp type options.
git-svn-id: file:///var/svn/chaste/projects/Angiogenesis@24005 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ src/vasculature/VasculatureGenerator.hpp
@@ -82,9 +82,9 @@ public:
     																		unsigned doublings_in_direction2,
     																		unsigned doublings_in_direction3 = 1);
 
-    boost::shared_ptr<CaVascularNetwork<DIM> > GenerateHexagonalNetwork(unsigned width,
-                    													unsigned height,
-                    													unsigned vessel_length);
+    boost::shared_ptr<CaVascularNetwork<DIM> > GenerateHexagonalNetwork(double width,
+                    													double height,
+                    													double vessel_length);
 
     /*
      * Creates a hexagonal repeating unit
